// recetas-crud/src/app/models/recipe.model.ts
// Esta interfaz representa la estructura de una receta.
// No es obligatorio incluir absolutamente todos los campos de la API,
// pero sí los que vamos a utilizar en la aplicación.

export interface Recipe {
  id: number;
  name: string;
  ingredients: string[];
  instructions: string[];
  prepTimeMinutes: number;
  cookTimeMinutes: number;
  servings: number;
  difficulty: string;
  cuisine: string;
  caloriesPerServing: number;
  tags: string[];
  image: string;
  rating: number;
  reviewCount: number;
  mealType: string[];
}